Zero initialize struct values to avoid uninitialized variables
Summary:
This came as a result of a misunderstanding by me of how zero initialization works in C++. I thought that since we had a constructor for the OutputMetricsData struct and used braced initialization, it would be zero initialized, but obviously this was wrong.

I've asked this question in the Learning C++ group to get a better explanation of why our reasoning was flawed in how this variable would be initialized.
https://fb.workplace.com/groups/474291069286180/permalink/5097561490292425

Reviewed By: peking2

Differential Revision: D26233103

fbshipit-source-id: f3043314a5bfa4894ee5e94bfc235aa0f038937b