Add Test/Control MatchCount to LiftMetrics
Summary:
# Context
For T83821858, we are adding a new metric: MatchCount. This measures the number of people who had an ad opportunity on Facebook and made a conversion at some point for the test/ control groups.

In order to add this metric to MPC, we need to modify `LiftMetrics` to hold these new metrics.

Reviewed By: gorel

Differential Revision: D26381594

fbshipit-source-id: fd30e38decb1d09093780a9862b7e09e1ac58d97