Add MatchCount metrics in PL
Summary:
# Context

For T83821858, we are adding a new metric: MatchCount. This measures the number of people who had an ad opportunity on Facebook and made a conversion at some point for the test/ control groups.

In, D26381594, this metric is added to the `LiftMetrics` sturct and further integrated within `MetricsMapper` and `EncryptedLiftMetrics`.

In this diff, MatchCount is calculated within `OutputMetrics.hpp`.

A new function `calculateMatchCount` determines if a opportunity is valid and that any conversion is made on this opportunity. A valid test/control match is when a person with an opportunity who made ANY nonzero conversion. Therefore we can just check if `opportunityTs > 0` and take a bitwise AND with the aggregate bitwise OR of `purchaseTS` (valid purchase) for all purchases a user made. This gets us a binary indication if a user is matched with any opportunity

Reviewed By: gorel

Differential Revision: D26428306

fbshipit-source-id: 88541c309036f4908d68cfdb18d57c112366e469