Create K-Anonymity Aggregator binary target
Summary:
Created a new K-Anonymity binary target as pl-aggregator-kanon.
Refactored KAnonymityLiftAggregationGame to pass in the threshold (instead of being hard coded to 100)
Completely split apart regular Aggregation vs KAnonymityAggregation (including test)

NOTE: This is essentially a duplication of a lot of Aggregator* however the intention is for this to move out in the future and not be dependent on the original Aggregator* code.

Reviewed By: gorel

Differential Revision: D27135468

fbshipit-source-id: d8d32aaaddaec0a7f65d05a0e8d7fd0223c173b2