Don't hide SumSq metrics since variance is hard
Summary:
I've been having a conversation with Erik and Paul and we've found that calculating LiftCI within MPC is much harder than expected. It's nearly impossible to calculate variance accurately in integer division land and moving to float division will add hours to the computation time (floats are ridiculously expensive in MPC).

As such, we agreed to instead expose SumSq metrics for now and work on a longer-term plan for calculating the CI within MPC.

Reviewed By: peking2

Differential Revision: D27272016

fbshipit-source-id: 9ffe194f02c57c1a28f8477e00550f4ce336509e