Aggregate cohorts even if lengths are not even between groups
Summary:
Imagine a scenario in which the following happens:
- An advertiser submits a dataset with three cohorts
- Shard0 processes cohorts 0 and 1
- Shard1 processes cohorts 0 and 2
- Shard2 processes cohorts 0, 1, and 2

In the current system, aggregator will fail because Shard0 won't have any data for cohort 2 and the lengths will be unequal. The solution is actually simple, though: We just pad Shard0's results with `0`. This works because XOR maintains the identity property s.t. `a ^ 0 = a`.

"What about Shard1 that was missing cohort 1?"
This case is actually handled in the Lift code itself: At the start of the code, we check the largest cohort_id in the dataset and during the calculator MPC, we loop from 0 to max(cohort_id), ensuring that gaps are properly filled.
NOTE: Long-term, an optimization is possible here if the advertiser allows us to see cohort_ids: this means we can do the bitmasking operation in the clear, which is *significantly* cheaper in MPC.

"Doesn't this leak something about groupings?"
Technically, yes. While running the lift calculator, the publisher may learn that cohorts K-N were not present in dataset X if the max cohort seen is K-1. In practice, this is likely to mean *very little* if advertisers have chosen their cohorts to be large populations instead of finely-tuned mini-groups. This is a problem that is not possible to be solve without coordination _outside_ the MPC and is currently seen as a low-risk problem.

Reviewed By: peking2

Differential Revision: D27561880

fbshipit-source-id: 940a5c7dc984aefccd0671b54249f9114b0cb445