Make Publisher and Partner roles consistent
Summary: The goal here is to make sure we use consistent values for the `publisher` and `partner` roles and have it align with the values we use for `Alice` and `Bob` in [EMP](https://github.com/emp-toolkit/emp-tool/blob/master/emp-tool/utils/constants.h#L13-L14) and even in other places within our code base. While, there is no hard requirement for these to be the same, it could cause hard to debug issues if we end up accidentally casting it from publisher to partner or vice versa.

Reviewed By: peking2

Differential Revision: D28008731

fbshipit-source-id: 740542a91f1008a74f42d856d8447ef12b391b7a