Move fbpcf to fbcode/ - Refactor test folder structure (#14)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/14

When we actually move fbpcf to fbcode root, we want to ensure there are no changes on GitHub to be made (because the diff that actually moves it will not be mirrored to GitHub).
If we move before doing this change, all the test files will be broken since the folder structure would have changed.

Next step: Rename pcf to fbpcf everywhere

Reviewed By: corbantek, peking2

Differential Revision: D29321161

fbshipit-source-id: 76a3ce20fb445dd4d96924699c678ccf388a5fdf