Fix the bug that containers can only get access to the buckets in the same region
Summary:
as title.

We got error:  ``error msg: Unable to parse ExceptionName: PermanentRedirect Message: The bucket you are attempting to access must be addressed using the specified endpoint. Please send all future requests to this endpoint. with address : 52.218.224.65 with address : 52.218.224.65 `` when containers and s3 buckets are not in the same regioin.

Differential Revision: D30172389

fbshipit-source-id: 025cbbab89d402e1e2b882ecda844bf521358c34