Fix the bug that containers can only get access to the buckets in the same region (#25)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/25

as title.

We got error:  ``error msg: Unable to parse ExceptionName: PermanentRedirect Message: The bucket you are attempting to access must be addressed using the specified endpoint. Please send all future requests to this endpoint. with address : 52.218.224.65 with address : 52.218.224.65 `` when containers and s3 buckets are not in the same regioin.

Reviewed By: peking2, ajaybhargavb

Differential Revision: D30172389

fbshipit-source-id: 9bdd60ac020307487d682a5fea64503fc58d7d2c