Documentation for preparation of v2 OSS release (#63)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/63

We are about to open source our v2 code on GitHub.  As part of that effort we are adding documentation for users and contributors of the project to better understand the problems we are trying to solve, the solutions that we designed to address the problems, and a guideline for how to migrate and start using the v2 APIs.

Differential Revision: D34702252

fbshipit-source-id: 13480518b76f056d54c9ad900b26933c22a385bb