Combine builds into one job (#92)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/92

I think the way the self hosted runner works is that it doesn't seem to preserve state between two jobs (since jobs from different workflows can be run in any order). So, we were seeing instances where the `sanity_check` job didn't find the fbpcf image.

This will fix it. The other option was to upload to the docker registry with a pre-release tag, but that would upload an image for every pre-release, which will use a lot of storage.

Reviewed By: wuman

Differential Revision: D34888385

fbshipit-source-id: 2636b192079b0d1573887b50257a94dd89456a8d