Use `dev` tag for coordinator image (#101)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/101

We need the coordinator to have Ajinkya's latest changes that introduce a PCF 2 stage flow

The `latest` tag is 9 days old. It does not have the changes.

In order to unblock our workflow, we can use the `dev` tag. It is stable because it's only pushed when the e2e tests for lift and attribution pass.

Reviewed By: jrodal98

Differential Revision: D34939006

fbshipit-source-id: bee998dcf8ac43a3da15bf80b88e06bbdd60bf3d