Track initialization metrics separately (#102)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/102

Many components in the MPC engine have a one-time initialization. Depending on how large of a sample size we use in our benchmarks, this can take up a significant amount of the overall runtime or network traffic. For this reason, I'm thinking it'd be useful to separate out the portion that's just for the initialization.

Reviewed By: adshastri

Differential Revision: D34906916

fbshipit-source-id: 9e02a44e8c99242801895ef11e95299129e443cb