Add a new API to Wirekeeper (#108)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/108

For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff simply add a new API to wirekeeper to get writable reference.

Reviewed By: elliottlawrence

Differential Revision: D34913258

fbshipit-source-id: eda97c01d0c2d5d4e0f871b3f75a4af266ff04be