Add an API to call rebatching gate in scheduler (#110)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/110

For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff adds the API to call this type of gates in the scheduler

Reviewed By: chualynn

Differential Revision: D34911752

fbshipit-source-id: 456612718029df3e6a93b384342a3c0f02a9f78a