Add a new API to Wirekeeper
Summary:
For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff simply add a new API to wirekeeper to get writable reference.

Differential Revision: D34913258

fbshipit-source-id: 6b63211b3d9884190ebe93e5d9ba336cc471512a