Add rebatch gate for lazy scheduler (#113)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/113

For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff add support of this type of gate in the gate keeper.

Reviewed By: elliottlawrence

Differential Revision: D34911728

fbshipit-source-id: 5d81813c9acdb7f100740c2abcd748788ab641ba