Add rebatch gate for lazy scheduler
Summary:
For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff add support of this type of gate in the gate keeper.

Differential Revision: D34911728

fbshipit-source-id: 8d03044903d56c7b35da88e514e01106cdd9af86