Add an API to call rebatching gate in scheduler
Summary:
For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff adds the API to call this type of gates in the scheduler

Differential Revision: D34911752

fbshipit-source-id: c7609fb9241f82812fb013d72c2619186704f4af