Add an API to call rebatching gate in scheduler (#110)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/110

For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff adds the API to call this type of gates in the scheduler

Reviewed By: chualynn

Differential Revision: D34911752

fbshipit-source-id: 2bcb0cb04e9c6a5977d9745db16b233926262e6e