Add a new API to Wirekeeper (#108)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/108

For UDP protocol, we need to add a new type of gate. Rebatching gate. This type of gate allows to break a batch of values into smaller batches or combine several batches into a larger one.

This diff simply add a new API to wirekeeper to get writable reference.

Reviewed By: elliottlawrence

Differential Revision: D34913258

fbshipit-source-id: d51d9c86ee896253e60ba1ea5d2db81eabc2577a