Add interface/class for composite tuple generation
Summary:
This diff adds the external interface for the tuple generator to support composite tuples. The caller will pass in an unordered map of composite tuple sizes to the number of those tuples requested.

I've also added a second method to perform both the composite and non-composite tuple generation together to reduce the number of round trips. I'm not actually sure how useful this will be though given the use of async buffers in the implementations so might end up removing it.

Reviewed By: RuiyuZhu

Differential Revision: D34807935

fbshipit-source-id: a4f45717e6f62e674c5e8d7266d1ecb270e7703e