Change composite tuple generator to use map instead of unordered_map (#129)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/129

Since `unordered_map` might have a different order of iteration depending on the platform, there is a potential for different parties to try to generate each tuple size in a different order. This will cause the tuples to not match up since the rcot results will be consumed out of order from the buffer.

I split this up from the next diff to make it easier to review.

Reviewed By: elliottlawrence

Differential Revision: D34980786

fbshipit-source-id: fdcb138b9b86de04895608d70599f28984a41a90