Implement BufferedWriter that wraps a child writer (#114)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/114

This diff adds a very simple buffered writer. Although this wasn't a requirement for supporting large PA/PL runs, it is good to have parity for buffered reader and buffered writer.

This class has a flush() function that writes the current buffer to the underlying file. When write is called, it first writes into the buffer and then flushes if there is no space

Reviewed By: elliottlawrence

Differential Revision: D34904877

fbshipit-source-id: cc6dd2056ee054155516abb15de25c2f0dd78478