Improve BufferedReaderTest to use parametrized tests and fix bug (#130)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/130

Make BufferedReaderTest use a parametrized test so that we can test various chunk sizes.

In addition, I discovered a bug where if a line ends at the last position of the buffer, eof() will not return false even if it was actually the end of the file. This is because the underlying `ifstream` only sets the `EOF` bit when you _actually try to read_ the end of the file, but nothing exists.

To fix this, we can load the next chunk when we reach the end of the buffer, which will attempt to read and then set the `EOF` bit.

Differential Revision: D35058349

fbshipit-source-id: db77cb5e0920f000cb47ac141c58d1302d8ddfa8