Use size_t in NpBaseObliviousTransfer
Summary:
We're seeing some random failures in the oblivious transfer benchmark.
https://www.internalfb.com/intern/servicelab/everstore_log/?handle=GD64rAjRdkkAm9gBANTW64QvwFdHbqMrAABn
  what():  Cant convert hex to point: 268857446: error:10067066:elliptic curve routines:ec_GFp_simple_oct2point:invalid encoding

I'm not entirely sure why, but I suspect there's an issue with sending/receiving data. I wonder if the size is overflowing, so perhaps using `size_t` instead will fix it. (I'm not able to repro this consistently, it only fails in ServiceLab sometimes.)

Reviewed By: RuiyuZhu

Differential Revision: D35118099

fbshipit-source-id: aec18f17ea43235cbe272d55c6264e8f922aab15