Change the behavior of BitString
Summary:
As title.
Current BitString type have some inconsistent behavior.

For example, when initialized by a batch of `std::vector<bool>`, it is expecting {the vector of first bits in all vecs, the vector of second bits in all vecs...} instead of "the first vector, the second vector". This doesn't agree with other batched types.

This diff changes this behavior to align with other batch types.

Differential Revision: D35122217

fbshipit-source-id: 3e64b628409c60b00c15d5f8d9e71a962cb561a1