Add a `Intp` type that can support any integer
Summary:
As title.
This diff adds a customized Intp type that can directly refect the behavoir of any integer (signed/unsigned with any reasonable width).
The point of this type is to allow using different width of integers in ORAM/shuffler/permuter for best performance.
Migrate to use it in ORAMs will be a BE task.

Reviewed By: chualynn

Differential Revision: D35098816

fbshipit-source-id: de0f9a93c31c1350567c2f45f4bfd881b27005ae