Add a `Intp` type that can support any integer (#154)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/154

As title.
This diff adds a customized Intp type that can directly refect the behavoir of any integer (signed/unsigned with any reasonable width).
The point of this type is to allow using different width of integers in ORAM/shuffler/permuter for best performance.
Migrate to use it in ORAMs will be a BE task.

Reviewed By: chualynn

Differential Revision: D35098816

fbshipit-source-id: 2ada15c01f65e1b44c795dbaf4c3cfb326b15e0b