Create overloaded constructor instead of default arguments (#159)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/159

Per discussion in D34216915, we should not make breaking API changes like this.

This diff creates the overloaded constructor. Note that this is a no-op because both versions continue to exist.

We are splitting up the two diffs because they are in different repositories (although in theory it is a safe change and shouldn't matter)

Reviewed By: chualynn

Differential Revision: D35198992

fbshipit-source-id: fb8aad9e857c261a9ba82cb7f93e6b9335b10610