Add benchmark for schedulers (#164)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/164

Looking for some feedback on this approach.

The scheduler API has 40 methods, and trying to test each one individually would be overkill I think. Plus, the implementation for a lot of the gates is very similar.

I figured the best thing to do would be to just create a bunch of dependent gates, that way we can test that the lazy scheduler properly batches things together more efficiently than the eager scheduler.

If anyone has suggestions for more thorough things to test, LMK.

Reviewed By: RuiyuZhu

Differential Revision: D35162133

fbshipit-source-id: 3d2cdfa5b01a2cd2a321c471eacc0377a8650ba4