Fix race condition when assigning version string (#166)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/166

Last night, a workflow failed with an interesting issue: https://fburl.com/xov78dbj

There were two commits landing at similar times, and the following sequence of events happened
1) `build_prerelease_image` of commit 1 was executed. it was assigned a version string of 1.0.3.
2) Then `private_attribution_e2e_test`/`lift` were executed for commit 1
3) The github executor decided to then execute `build_prerelease_image` for commit 2 (it does that a lot, the granularity is at the job level not the workflow level)
4) `build_prerelease_image` for commit 2 examined the history of the repo and also assigned a version string of 1.0.3 because there was never a tag of 1.0.3.
5) commit 1 ran `tag_and_push` and created a release of version 1.0.3.
6) commit 2 tried to release 1.0.3 and got an error

This diff moves the version string generation into the final tag and push stage so it will be atomic.

The only functionality we lose here is that there is no increment assigned for prereleases. This is fine for now since it's totally unused. We can revisit this.

Reviewed By: wuman

Differential Revision: D35255590

fbshipit-source-id: a6ace1432096a824baa2669ab8bfa32cda1c836f