Make entire docker-publish workflow atomic (#178)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/178

There is actually one more concurrency issue. Consider the following scenario:

1) Commit A runs `build_prerelease_image` and uploads the prerelease image to `test_registry:test_image`
2) Commit A gets preempted by Github, and now Commit B does the same thing.
3) `test_registry:test_image` now contains Commit B's image, and Commit A's image is lost
4) when both Commit A's and Commit B's `private_lift_e2e_test` runs, it uses Commit B's image

Commit A will never get tested

Unfortunately, I think this means we have to make the entire workflow atomic. We only have one self hosted runner so it doesn't really change much for us.

Interestingly, fbpcs has always done it this way, so I guess they had more foresight than me :)

The proper solution would be to see if we can specify the image tag when running on the PCE, but at the moment that seems impossible since the image location is configured in the AWS console

Reviewed By: wuman

Differential Revision: D35405964

fbshipit-source-id: 0add9c71ebdbdd9a2c87e27fa76ee29d471f6775