Migrate PID sharder to use new BufferedReader (#195)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/195

X-link: https://github.com/facebookresearch/fbpcs/pull/829

This diff makes the change in production to use the new Buffered Reader. Google doc for context is linked below (internal only).

See test plan for full details of how this was tested.

There is one workaround about `SIGPIPE` that is explained with an in-line comment.

Differential Revision: D35729872

fbshipit-source-id: 0ea0bd1afb98b71657f8f9c229a1f48d77c05ae2