Create workflow to publish new dependency for emp when we upgrade
Summary:
# Context
FBPCF builds have a caching mechanism. When we build our docker image, we have cached the images for emp, aws, gcp, and folly so that we don't have to rebuild them every time. However, when we upgrade a dependency (like we did with emp a few months ago), this image is stale and gets rebuilt every time we build fbpcf. So, we need to create a process to rebuild the cached image.

The solution here, IMO, is to create 4 workflows, one for each dependency, that accepts a version number as an argument and will build and publish a dependent image for that version. It will only be called on `workflow_dispatch` (i.e. manually), not at `push` or `pull_request` time.

Each diff sets up a workflow for the respective based on the code here: https://fburl.com/code/zyma3fzr

# This Diff
Workflow for emp

# This Stack
1. **Workflow for emp**
2. Workflow for aws
3. Workflow for gcp
4. Workflow for folly

Differential Revision: D36284480

fbshipit-source-id: f1085e19d91951391f736459da8cf7437b375ee9