Create workflow to publish new dependency for emp when we upgrade (#204)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/204

# Context
FBPCF builds have a caching mechanism. When we build our docker image, we have cached the images for emp, aws, gcp, and folly (https://github.com/orgs/facebookresearch/packages?repo_name=fbpcf) so that we don't have to rebuild them every time. However, when we upgrade a dependency (like we did with emp a few months ago), this image is stale and gets rebuilt every time we build fbpcf. So, we need to create a process to rebuild the cached image.

The solution here, IMO, is to create 4 workflows, one for each dependency, that accepts a version number as an argument and will build and publish a dependent image for that version. It will only be called on `workflow_dispatch` (i.e. manually), not at `push` or `pull_request` time.

Each diff sets up a workflow for the respective based on the code here: https://fburl.com/code/zyma3fzr

# This Diff
Workflow for emp

# This Stack
1. **Workflow for emp**
2. Workflow for aws
3. Workflow for gcp
4. Workflow for folly

Reviewed By: RuiyuZhu

Differential Revision: D36284480

fbshipit-source-id: 515bc1a7d7cdeccd8a2e6a5bd5d76abe9d3b4474