Update the cloud file writer to use shared S3Client (#214)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/214

Aditya recently found an issue that when we have 169 shards for uploading files, it will fail at 169 shards. Similar to this issue https://github.com/aws/aws-sdk-cpp/issues/1920. After Aditya talking with AWS team, the current solution is to use a shared S3Clinet instead of creating one per each request. In the meanwhile, AWS is working on finding the root cause. This diff is to make the S3Client a singleton and use the same client among all requests.

Reviewed By: gorel, gitfish77

Differential Revision: D36727729

fbshipit-source-id: e4d62ce56b5da616d5af0bb7e7da50a013089986