Change from reference to unique_ptr, and make the change push safe- Step 1
Summary: Create a duplicate version of the BufferedReader called BufferedReaderCopy. It can have exactly the same code, just copy pasted. Let's also duplicate the BufferedReaderTest just as a sanity check. Get that diff accepted, and land it. You will need to modify the TARGETS file to include the .h file(s) in headers and the .cpp file(s) in srcs.

Reviewed By: adshastri

Differential Revision: D37145220

fbshipit-source-id: d1a3ee8c38642914f6893b5b43ce875c8baa6913