Use singleton S3Client for reads
Summary:
# What
Previously, we noticed a bug during multipart uploads that multiple concurrent S3Clients would result in failed writes. So, we created a singleton S3Client to share between all write operations. This diff does it for reads.

# Why
It is a better architecture, since we do not need multiple S3Client objects.

Reviewed By: wenhaizhu

Differential Revision: D37557580

fbshipit-source-id: 322c7d742343f5016a42d8e7eb27d0f86ba01333