Use singleton S3Client for reads (#254)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/254

# What
Previously, we noticed a bug during multipart uploads that multiple concurrent S3Clients would result in failed writes. So, we created a singleton S3Client to share between all write operations. This diff does it for reads.

# Why
It is a better architecture, since we do not need multiple S3Client objects.

Reviewed By: wenhaizhu

Differential Revision: D37557580

fbshipit-source-id: 30f7044a51f7be35459d5354aa5bef91a413a404