Add an adapter for boolean type (#255)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/255

Add a MPC adapter for a boolean type with an implementation of oblivious swap for a batched type `Bit`. The boolean adapter allows us to run the shuffler and permuter for a boolean type of secret values.

It will be used as a building block for a shuffler-based compactor which requires boolean type of values to be shuffled.

Reviewed By: RuiyuZhu

Differential Revision: D37476874

fbshipit-source-id: ed0375c0c213fb2008d4f7576b4a3951cb742db1