Refactor dummy permuter to use MPC adapters (#259)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/259

Use MPC adapters when opening and processing secret values in dummy permuter implementation.

Details: In the current implementation of dummy permuter, a template type ```T``` is limited to Bit, Int and BitString as it has a dependency on the method ```OpenToParty``` and the constructor for processing secret inputs. Thus, to make ```T``` more general, we use MPC adapters as proxies to call those functions.

- We make small changes to all existing dependent unit tests to reflect the modifications to the dummy permuter.

Reviewed By: chualynn

Differential Revision: D37687008

fbshipit-source-id: abe3d2e245d6a9fbf66350508153e31bf5c5802f