Refactor dummy permuter to use MPC adapters (#1243)
Summary:
X-link: https://github.com/facebookresearch/fbpcs/pull/1243

Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/259

Use MPC adapters when opening and processing secret values in dummy permuter implementation.

Details: In the current implementation of dummy permuter, a template type ```T``` is limited to Bit, Int and BitString as it has a dependency on the method ```OpenToParty``` and the constructor for processing secret inputs. Thus, to make ```T``` more general, we use MPC adapters as proxies to call those functions.

- We make small changes to all existing dependent unit tests to reflect the modifications to the dummy permuter.

Reviewed By: chualynn

Differential Revision: D37687008

fbshipit-source-id: 8eb07886a5e869862181b5e3c097f74bf931ab12