Input Reader for edit distance app (#258)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/258

Adds a input reader to the edit distance app which can process the input files from the data generator. The input reader class takes two paths to the constructor, the parameters path, and one of the input files (i.e. player 1 or player 2)

I copied the Csv file reader implementation from https://fburl.com/code/eiinihwe in the FBPCS repo since we shouldn't depend on it and modified it a bit to be simpler.

The constructor loads the input lines into its member variables and increments the number of rows processed. The variables for the other party remain empty and will be up to the other components to handle properly.

Reviewed By: chualynn

Differential Revision: D37627524

fbshipit-source-id: a3e0bdc4984590c7f0e64e6298968b2bb14e1e34