InputProcessor (#265)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/265

The InputProcessor component takes the input files from the previous diff that have been loaded into memory and converts the inputs into secret share format.

The constructor takes a `EditDistanceInputReader` and integer indicating the players role. Construction of the object will

1. Validate consistency of input rows as well as params (threshold, delete, insert cost)
2. Instantiate private values (words, senderMessages, guesses) into the class.

This object can then be directly used to compute the results of the game within MPC.

Reviewed By: chualynn

Differential Revision: D37769029

fbshipit-source-id: 94efc7e23596345259e5e2590da8a0c65857eb99