Implement private size (#281)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/281

This is needed for the EditDistanceCalculator when determining the bounds of the letters as well as the receiver messages.

The AsciiString class represents the end of the string by a padded set of null characters. We are able to determine the size of the string by scanning through each character until a null char is found.

This method returns a Secret Integer representing the real length of the string.

Reviewed By: RuiyuZhu

Differential Revision: D37871747

fbshipit-source-id: 1563da911f3f4a04dd9c7b0308ae501e0ac70df4