Implement oblivious index operation
Summary:
This is needed for the concat function in AsciiString when accessing a character in the AsciiString while the index is secret.

Overloaded the [ ] operator so the index could be a secret Int object. Iterate through the AsciiString and assign the return value to be the current character if we are at the secret index. Return a signed secret Int object.

Differential Revision: D37892641

fbshipit-source-id: 128c531dccd6da5f6df824770da42994158bff01