Implement Integer Casting
Summary:
Implements integer casting in MPC. I followed the rules outlined here https://stackoverflow.com/questions/27889213/c-integer-downcast.

> For unsigned to unsigned types, it adds sufficient zero most-significant bytes; this always preserves the value.
For signed to signed types, it sign-extends the the source type (i.e. packs the new byte(s) with bits equal to the sign bit of the source integer); this always preserves the value, positive or negative
For unsigned to signed types, it effectively adds sufficient zero-most significant bytes; this always preserves the value, as in the nature of an upcast, there are always more bits in the destination, so there is always space for an extra sign 'bit'
For signed to unsigned types, it sign-extends, then casts; this cannot always preserve the value, as a negative value cannot be represented.

This is needed to cast the ascii chars to a higher value so they can be summed up into a distance that may exceed the Ascii character range.

Reviewed By: RuiyuZhu

Differential Revision: D37871978

fbshipit-source-id: 498d7d33625503d591973589236e87beb3ecfcc7