Revert diff D37557580 (#289)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/289

We wanted to use a singleton S3Client for S3FileReader also, but unfortunately this singleton is shared between S3FileReader and S3FileUploader. So if the reader and uploader are touching files in different region, it will throw error "The bucket you are attempting to access must be addressed using the specified endpoint. Please send all future requests to this endpoint.". The reason is that the reader will initilize S3Client with one region (e.g. us-west-2), and then the uploader will reuse it to upload file to another region (e.g. us-west-1).

Reviewed By: anthonyzhang25, adshastri, gitfish77

Differential Revision: D37977599

fbshipit-source-id: ed6cf3a9f7380c050d7a9bc3a491a09d845080b8