Delete PIDInstanceRepository (#288)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/288

PIDInstanceRepository is used by the old PIDService which was replace by the new PID stage services (D37619253). In this diff, we delete all PIDInstanceRepository related code.
And because we updated the repo configs for mpc_aem/experimentation_platform.  EP also used manifold config files to run continuous measurements.  The same changes should be included in continuous measurement as well. We assign the task (T126244336) to ajinkya-ghonge

# Changelog
[Removed] - Delete PIDInstanceRepository

Reviewed By: jrodal98, joe1234wu

Differential Revision: D37874084

fbshipit-source-id: dc6b2a2d32dc3f0626d92410102a72d9ce413e99