Implement edit distance business Logic (#286)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/286

Implements the core logic to calculate the edit distance. The algorithm is slightly complicated by the fact that the size of the strings is unknown, so the solution matrix must account for that.

{F753583602}

I drew the following diagram as an example. The rows direction has the letters of the word, and the columns are the letters of the guess that we are trying to change into the word.

Moving up means we are inserting the letter into the guess at that position of the word. Moving left means we are deleting the letter of the guess at the current position. Moving diagonally means replacement of the letter.

Once we are out of bounds, the logic depends on which quadrant we are in. Assuming the quadrants are numbered
1 2
3 4
quadrant 2 indicates we should copy the value to the left, quadrant 3 is copying the value from above, and quadrant 4 is copying the value diagonally up and left. This guarantees that the correct solution in the position of the last letter of guess and last letter of word will surround quadrant 4, and then be copied into the last position.

Reviewed By: RuiyuZhu

Differential Revision: D37792933

fbshipit-source-id: 0291240247b813147bf2c496e0b2c1db1cd3f949