Serialize EditDistance game results (#305)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/305

Adds a helper method to serialize the results of the EditDistance game. It will extract the integer and string shares of the `receiverMessages` and `editDistances` into a folly::dynamic object and then serialize with `folly::json`.

This will be used to write the results to disk and then validate in a later stage.

Differential Revision: D38217636

fbshipit-source-id: 6ab1b14924b67deec477aed9a65088d81176b7ec