Add EditDistanceResults deserialization helper class for shares (#307)
Summary:
This diff adds deserialization logic for secret shares of the edit distance results. These are obtained by the toJSON function added in D38217636. I updated the test that I added in that diff which combines all the shares to use this helper class, and will use it in the next diff as well.

Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/307

Differential Revision: D38231945

fbshipit-source-id: c9cf05ab67090738602a1f166b555ac83a9b87ae