Fix broken AsciiString Test (#317)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/317

Fixing the broken test in T127614039. Seems the changes in D38231945 (https://github.com/facebookresearch/fbpcf/commit/416ab36f0627383034be53e3a68c76bbbddf9dbc) had a bug. I was able to re-create this locally with a stress test. Turns out when one of the character "shares" is set to 0, calling getValue() will not include it in the string. This means the string form of the extracted signed characters will be missing a letter. When we call parse JSON on this and try to XOR the shares, they are misaligned and the answer is bogus.

Reviewed By: RuiyuZhu

Differential Revision: D38334286

fbshipit-source-id: 6fff69671b0a97c7d654ff2c7566dbe4aa33a067