Clang Build for Edit Distance application + Github Action (#322)
Summary:
This diff adds two binaries to the docker image (`edit_distance_runner` and `edit_distance_validator`) which run the scripts added in the previous two diffs. It also adds a new stage in the github actions to run edit distance on each pull request to the Github repository. The files are downloaded from AWS and results saved to an `outputs` folder which the validator will then check.

Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/322

Differential Revision: D38448981

fbshipit-source-id: ac7ddb14b7b203bf44ff4c70c91d20ceb2518022