Add Validator class (#318)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/318

Adds the validation step of the Edit distance application. The constructor accepts the output file paths of the shares, and then recovers them using the `EditDistanceResults` class. It will then read the ground truth results file and compare the end result.

Side note: The EditDistanceResults class technically surpasses extractIntShare() and openToParty methods which would make more sense as a reveal stage. For now I think this is fine, but if we have different MPC backends to test then we should use the native MPC API's to "reveal" the results to one of the players. I will put this on our backlog as a BE.

Reviewed By: RuiyuZhu

Differential Revision: D38260494

fbshipit-source-id: ed599bb4255f8308cecfe832060d0528f580df77