Separate default argument into overloaded constructor
Summary:
# Context
We want to have more intelligent defaults for the IO APIs. Cloud related upload/downlaod should have a larger buffer size to minimize IO operations, whereas local io should be smaller.

# This Diff
We are just creating a separate constructor that we will modify in the future to be more intelligent

# This Stack
1. **Create separate constructor without chunkSize**
2. Add a filepath member variable to IWriter and IReader
3. Create util functions to determine buffer size
4. Use the new util functions
5. Delete all other references to determining chunk sizes

Differential Revision: D38873747

fbshipit-source-id: e8388b8240b14e43a4355bf6b24b6e89f9cd645b