Add getFilePath() function to all readers and writers (#360)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/360

# Context
We want to have more intelligent defaults for the IO APIs. Cloud related upload/downlaod should have a larger buffer size to minimize IO operations, whereas local io should be smaller.

# This Diff
We are exposing a `getFilePath` function in all writers and readers so that we can then use it in the future for determining a default buffer size

# This Stack
1. Create separate constructor without chunkSize
2. **Add a filepath member variable to IWriter and IReader**
3. Create util functions to determine buffer size
4. Use the new util functions
5. Delete all other references to determining chunk sizes

Reviewed By: chualynn

Differential Revision: D38873574

fbshipit-source-id: 470d8161bf0d570f6f60b05c4eede8fbfac4cf36