Fix Compactor Benchmark (#374)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/374

I tried running this a couple times, it sometimes works and sporadically breaks. I noticed the method setCompactorFactory() is called twice which will set the member variables twice. Not sure why this was needed so I removed it and got rid of the base class (right now we only need one base factory).

These factories can be created before the scheduler is created, and then the compactor factory can be created after.

Reviewed By: chualynn

Differential Revision: D39043195

fbshipit-source-id: b2bf89b7cc9b199acc343193eb35ad2e82663912