Refactor IBidirectionObliviousTransfer (#352)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcf/pull/352

Delete the template class in IBidirectionOT and split the bidirectionOT function in IBidirectionOT to two functions, one for bool and one for uint64_t.

This update allows easier implementation of the integer product share generation in productShareGenerator, because productShareGenerator wouldn't need a template class to specify the type of BidirectionOT it will use.

Reviewed By: RuiyuZhu

Differential Revision: D38850680

fbshipit-source-id: fb262df1e40eb05652628fff4b4b3c2669e0c26f